poc/requirements.txt
streamlit
pandas
kaggle
Pillow

